# 4/openapi.yml
openapi: 3.0.3
info:
  title: Unified E-Commerce API
  description: Combined API for Order Management and Authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Unified API Gateway
paths:
...
  /auth/users/me:
    get:
      summary: Get Current User
      operationId: get_current_user
      tags:
        - AuthService
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum: [client, admin, courier, warehouse_manager]
      description: User role in the system

    # Auth Service
...
    User:
      type: object
      required: [user_id, username]
      properties:
        user_id:
          type: integer
        username:
          type: string
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
...